Drop support for Python 2 from the build process
Update `setup.cfg` so it does not build a Python 2 wheel and update
`setup.py` so that package metadata will correctly express Python 3+
only support. Due to the way this package is written trying to install
from the source tarball on Python 2 will result in an exception due to
attempted import of the UTF-8 version.py file.